pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "better-drinking-bird"
version = "0.1.0"
description = "Supervisor for AI coding agents - keeps them on task like Homer's drinking bird"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "cognaterra" }
]
keywords = ["ai", "coding", "agent", "supervisor", "claude", "cursor", "copilot"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "httpx>=0.25",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1",
]

[project.scripts]
bdb = "drinkingbird.cli:main"

[project.urls]
Homepage = "https://github.com/cognaterra/better-drinking-bird"
Documentation = "https://github.com/cognaterra/better-drinking-bird#readme"
Repository = "https://github.com/cognaterra/better-drinking-bird"
Issues = "https://github.com/cognaterra/better-drinking-bird/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/drinkingbird"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
